+++ arrow-libs/core/arrow-core-data/src/test/kotlin/arrow/syntax/function/MemoizationTest.kt
@@ -1,5 +1,6 @@
 package arrow.syntax.function
 
+import arrow.core.memoize
 import io.kotlintest.properties.forAll
 import io.kotlintest.shouldBe
 import io.kotlintest.specs.FreeSpec
