+++ arrow-libs/core/arrow-core-data/src/test/kotlin/arrow/syntax/test/CollectionsSyntaxTests.kt
@@ -16,9 +16,8 @@
 
 package arrow.syntax.test
 
-import arrow.syntax.collections.destructured
-import arrow.syntax.collections.prependTo
-import arrow.syntax.collections.tail
+import arrow.core.prependTo
+import arrow.core.tail
 import arrow.core.test.UnitSpec
 import io.kotlintest.shouldBe
 
@@ -35,7 +34,7 @@ class CollectionsSyntaxTests : UnitSpec() {
     }
 
     "destructured" {
-      val (head, tail) = listOf(1, 2, 3).destructured()
+      val (head, tail) = listOf(1, 2, 3).let { it.first() to it.tail() }
       head shouldBe 1
       tail shouldBe listOf(2, 3)
     }
