+++ arrow-libs/core/arrow-core-data/src/test/kotlin/arrow/syntax/test/FunctionSyntaxTest.kt
@@ -1,24 +1,22 @@
 package arrow.syntax.test
 
-import arrow.syntax.function.andThen
-import arrow.syntax.function.bind
-import arrow.syntax.function.complement
-import arrow.syntax.function.compose
-import arrow.syntax.function.curried
-import arrow.syntax.function.forwardCompose
-import arrow.syntax.function.invoke
-import arrow.syntax.function.memoize
-import arrow.syntax.function.paired
-import arrow.syntax.function.partially1
-import arrow.syntax.function.partially2
-import arrow.syntax.function.partially3
-import arrow.syntax.function.partially4
-import arrow.syntax.function.partially5
-import arrow.syntax.function.reverse
-import arrow.syntax.function.tripled
-import arrow.syntax.function.uncurried
-import arrow.syntax.function.unpaired
-import arrow.syntax.function.untripled
+import arrow.core.andThen
+import arrow.core.complement
+import arrow.core.compose
+import arrow.core.curried
+import arrow.core.forwardCompose
+import arrow.core.memoize
+import arrow.core.paired
+import arrow.core.partially1
+import arrow.core.partially2
+import arrow.core.partially3
+import arrow.core.partially4
+import arrow.core.partially5
+import arrow.core.reverse
+import arrow.core.tripled
+import arrow.core.uncurried
+import arrow.core.unpaired
+import arrow.core.untripled
 import arrow.core.test.UnitSpec
 import io.kotlintest.shouldBe
 import java.util.Random
@@ -179,7 +177,7 @@ class FunctionSyntaxTest : UnitSpec() {
     }
 
     "testReverse" {
-      val j: (String, List<String>) -> List<String> = f(p2 = 1)
+      val j: (String, List<String>) -> List<String> = f.partially2(1)
       j("x", listOf("a", "b", "c")) shouldBe j.reverse()(listOf("a", "b", "c"), "x")
     }
 
@@ -198,12 +196,12 @@ class FunctionSyntaxTest : UnitSpec() {
 
     "partials" {
       val sum5ints = { a: Int, b: Int, c: Int, d: Int, e: Int -> a + b + c + d + e }
-      val sum4intsTo10: (Int, Int, Int, Int) -> Int = sum5ints(p5 = 10)
-      val sum3intsTo15: (Int, Int, Int) -> Int = sum4intsTo10(p4 = 5)
-      val sum2intsTo17: (Int, Int) -> Int = sum3intsTo15(p3 = 2)
+      val sum4intsTo10: (Int, Int, Int, Int) -> Int = sum5ints.partially5(10)
+      val sum3intsTo15: (Int, Int, Int) -> Int = sum4intsTo10.partially4(5)
+      val sum2intsTo17: (Int, Int) -> Int = sum3intsTo15.partially3(2)
       sum2intsTo17(1, 2) shouldBe 20
       val prefixAndPostfix = { prefix: String, x: String, postfix: String -> "$prefix$x$postfix" }
-      val helloX: (String) -> String = prefixAndPostfix(p1 = "Hello, ")(p2 = "!")
+      val helloX: (String) -> String = prefixAndPostfix.partially1("Hello, ").partially2("!")
       helloX("Arrow") shouldBe "Hello, Arrow!"
     }
 
@@ -213,7 +211,7 @@ class FunctionSyntaxTest : UnitSpec() {
         i += a
       }
 
-      val binded = ::inc.bind(5)
+      val binded = ::inc.partially1(5)
       i shouldBe 0
       binded()
       i shouldBe 5
